requirements.txt
hs2p>=2.3.0,<3
omegaconf>=2.3.0
h5py
matplotlib
numpy<2
pandas
pillow
cucim
rich
torch
torchvision
transformers
tqdm
wandb
wholeslidedata<0.0.16
einops
timm
